Migrate CRC32 implementation to private implementation
Removes the template-based function generation.
